Fix contains() check
Eclipse is flagging a type mismatch, fix that up.

+++ applications/southbound-cli/src/main/java/org/opendaylight/openflowplugin/applications/southboundcli/DefaultReconcileService.java
@@ -144,7 +144,7 @@ public final class DefaultReconcileService implements Reconcile, ReconcileServic
         }
 
         final var allNodes = getAllNodes();
-        final var unresolvedNodes = nodes.stream().filter(node -> !allNodes.contains(node))
+        final var unresolvedNodes = nodes.stream().filter(node -> !allNodes.contains(node.longValue()))
             .collect(Collectors.toList());
         if (!unresolvedNodes.isEmpty()) {
             return buildErrorResponse("Error executing command reconcile. "
